+++ src/main/java/com/tallerwebi/dominio/ServicioJugador.java
@@ -1,7 +1,9 @@
 package com.tallerwebi.dominio;
 
 public interface ServicioJugador {
-    Boolean equipaCaniaAPersonaje(Jugador jugador, Integer idObjeto);
+
+
+    Boolean equipaCaniaAPersonaje(Jugador jugador, Long idObjeto);
 
     Jugador inicializarJugador();
 }
