+++ src/main/java/com/tallerwebi/dominio/ServicioJugadorImpl.java
@@ -2,8 +2,6 @@ package com.tallerwebi.dominio;
 
 import com.tallerwebi.dominio.excepcion.ObjetoInexistenteException;
 import com.tallerwebi.dominio.excepcion.ParametroInvalidoException;
-import com.tallerwebi.infraestructura.RepositorioJugadorImpl;
-import com.tallerwebi.infraestructura.RepositorioObjetoImpl;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
@@ -22,7 +20,7 @@ public class ServicioJugadorImpl implements ServicioJugador {
     }
 
     @Override
-    public Boolean equipaCaniaAPersonaje(Jugador jugador, Integer idObjeto) {
+    public Boolean equipaCaniaAPersonaje(Jugador jugador, Long idObjeto) {
 
         if(jugador == null || idObjeto == null){
             throw new ParametroInvalidoException("El jugador o el id del objeto no puede ser nulo");
